requirements.txt
google-cloud-storage==1.31.0
Flask~=2.0.1
flask-SocketIO==5.1.1
Flask-cors==4.0.1
pdfplumber==0.5.28
Pillow~=10.4.0
openai==1.13.3
requests==2.27.1
gunicorn~=22.0.0
langdetect
pdf2image==1.16.0
pytesseract==0.3.9
python-docx
PyMuPDF==1.24.3
gunicorn
python-dotenv~=1.0.1
Jinja2~=3.1.4
Werkzeug==2.2.2